// src/app/interfaces/Sort.ts
export interface Sort {
  belorussian: {
    countAll: number;
    countCurr: number;
  };
  foreign: {
    countAll: number;
    countCurr: number;
  };
  computer: {
    countAll: number;
    countCurr: number;
  };
  scientific: {
    countAll: number;
    countCurr: number;
  };
  artistic: {
    countAll: number;
    countCurr: number;
  };
  business: {
    countAll: number;
    countCurr: number;
  };
  children: {
    countAll: number;
    countCurr: number;
  };
  stok: {
    countAll: number;
    countCurr: number;
    countOrderAll: number;
    countOrderCurr: number;
    countNoAll: number;
    countNoCurr: number;
  };
  type: {
    countSoftAll: number;
    countSoftCurr: number;
    countHardAll: number;
    countHardCurr: number;
    countIntAll: number;
    countIntCurr: number;
  };
}